Added -lowmem mode in command line parameters. It enables low memory consumption mode but disables possibility to adjust image parameters after rendering. Memory consumption is 4 times less. The program doesn't use complexImage and CompileImage() function. Instead of this the image is compiled directly druring rendering image pixels. Most of shaders and effects works properly. Only reflections are not vissible when ambient occlusion is rendered using SSAO. Lowmem mode works in noGUI mode and also when GUI is used.

+++ mandelbulber/files.cpp
@@ -430,13 +430,13 @@ bool FileIfExist(char *filename)
 void WriteLog(const char *text)
 {
 	FILE *logfile = fopen(logfileName, "a");
-	fprintf(logfile, "%d: %s\n", (int) clock(), text);
+	fprintf(logfile, "%ld: %s\n", (unsigned long int) clock(), text);
 	fclose(logfile);
 }
 
 void WriteLogDouble(const char *text, double value)
 {
 	FILE *logfile = fopen(logfileName, "a");
-	fprintf(logfile, "%d: %s, value = %g\n", (int) clock(), text, value);
+	fprintf(logfile, "%ld: %s, value = %g\n", (unsigned long int) clock(), text, value);
 	fclose(logfile);
 }
